[PM-17011] Move network managers to `network` package
Moves `NetworkConfigManager` and `NetworkConnectionManager` to the `network` subpackage.

This change is for better organization of the project. There are no functional changes.
