+++ app/src/main/java/com/x8bit/bitwarden/data/platform/manager/network/NetworkConfigManager.kt
@@ -1,4 +1,4 @@
-package com.x8bit.bitwarden.data.platform.manager
+package com.x8bit.bitwarden.data.platform.manager.network
 
 /**
  * Responsible for managing the active configuration of the network layer.
