+++ app/src/test/java/com/x8bit/bitwarden/data/platform/manager/util/FakeNetworkConnectionManager.kt
@@ -1,6 +1,6 @@
 package com.x8bit.bitwarden.data.platform.manager.util
 
-import com.x8bit.bitwarden.data.platform.manager.NetworkConnectionManager
+import com.x8bit.bitwarden.data.platform.manager.network.NetworkConnectionManager
 
 class FakeNetworkConnectionManager(
     override val isNetworkConnected: Boolean,
